README: add Redox Beta to supported chargers
Add Redox Beta to list of supported chargers. Device is using the
typical "imaxb6-clone" configuration.

Only tested on newer revision with yellow soldermask.
Either soldering to CPU pins or replacing it with hot-air is required -
- it was possible to read out stock firmware.

There is an older revision of the charger with red soldermask and
programming port, but I haven't had a chance to get my hands on one,
yet.