[#337] Fix: Current (1A) is flowing when entering calibration menus.
Previously we tried to discharge the output capacitor before
starting anything including calibration or charging.

Output capacitor discharge was introduced with led support
and unnecessarily extended to everything.
It seems that it is impossible to discharge only the capacitor
if a battery is connected to the output, because there will always
be current flowing through the transistors diode at "batt (-)"